pyproject.toml
[project]
name = "multiagent-bot-sim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "watchdog>=6.0",
    "python-telegram-bot>=21.0",
]
